Add our template Dockerfile
This is mostly based on the duplicated Dockerfiles we had before---key
difference here is that we pull out the shell scripts and just put
them in the Dockerfile.  Since we have this as a master, we can just
do it all here; the procedure for setup seems to be the same across
all versions.

Co-authored-by: Hawken Rives <hawkrives@gmail.com>