Update README.md
I realize this may seem redundant to propose, but since the info in this readme is what's put into the dockerhub page for the images (https://hub.docker.com/r/lucee/lucee), there's nothing there that points folks to this github repo, so it seems a simple change to consider. Hope you would agree. If not, that's fine.